Added bigint for diagnosis migration.

+++ db/migrate/20190611094151_create_diagnosis.rb
@@ -2,7 +2,7 @@ class CreateDiagnosis < ActiveRecord::Migration[5.2]
   def change
     create_table :diagnosis, :primary_key => :diagnosis_id do |t|
     	t.bigint    :encounter_id
-      t.integer    :concept_id
+      t.bigint    :concept_id
     	t.boolean    :primary_diagnosis
       t.boolean    :secondary_diagnosis
       t.boolean   :voided, null: false, default: 0
